edit 3 on task 11

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -1,4 +1,3 @@
-#include "main.h"
 #include <stdio.h>
 
 /**
@@ -9,8 +8,8 @@
 
 int main(void)
 {
-	int x = 612852475143;
-	int y;
+	long int x = 612852475143;
+	long int y;
 
 	for (y = 2; y < x; y++)
 	{
